I don't get it. Might be model issue maybe? Unsure. Will bug Paul l8r.
Applying the crash test controller to the in-game model didn't actually fix the issue of the animations not working. It works in the test scene, but not in-game. It might likely be a blend tree issue, but it may be an issue with the model. Incorporating the animation controller inside of the player script might fix the issue maybe, but I will have to bug Paul when he's awake and around to really get to the bottom of this as the internet is NOT helping.