Changed anim-controller of Octo, modified blend tree.
Based on the short meeting with Paul yesterday, I switched the blend mode from 2D Cartesian to 1D, and removed the second movement motion that would've taken care of the backwards movement. Still waiting on Paul to help out as he said he'd look over some stuff today and get back to me. Will be working on animations in the meantime.